seperate RoleManager and DefaultRoleManager to different files
